backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
openai>=1.12.0
sqlalchemy>=2.0.0
aiofiles>=23.2.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
numpy>=1.24.0
chromadb>=0.4.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
# PostgreSQL driver for production deployments (Render, etc.)
psycopg2-binary>=2.9.9
# Optional: sentence-transformers (may not work with Python 3.13)
# If you get errors, you can use OpenAI embeddings instead
# sentence-transformers>=2.3.0

